Remove Jira deploy date check
With discussion from Colin and Nick, we are deciding to not worry about
deploy date in PreDeployChecker. The ideal state would be to have the
deploy_date automatically set in Jira when moving a ticket from Ready To
Deploy to PostDeploy/Closed. It was deemed that the removal of the check
can  be worked before the automation of this is complete